+++ CHANGES.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.4.0 - 2024-01-27
+* Implemented `--directory` option to rename files with the
+  directory name as prefix (@manuerwin)
+
 ## 1.3.0 - 2022-05-25
 * Update GoPro filename patterns (@jaymoid) and a link to related
   documentation (@d-bucur)